requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.1
django-filter==2.1.0
django-heroku==0.3.1
geographiclib==1.49
geopy==1.20.0
gunicorn==19.9.0
idna==2.8
pipenv==2018.11.26
psycopg2==2.8.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
virtualenv==16.7.1
virtualenv-clone==0.5.3
whitenoise==4.1.2
win-inet-pton==1.1.0
wincertstore==0.2
